ESP32 WROOM - Send Email on Button Press
This project uses the ESP32 WROOM microcontroller to send an email when a button is pressed. The ESP32 connects to a WiFi network and uses an SMTP server to send the message. 


Possible Expansions
This project can be expanded by replacing the button with a motion sensor or another type of sensor. This way, ESP32 can automatically send an email when movement is detected.